+++ app/src/main/java/com/jakewharton/telecine/TelecineActivity.java
@@ -4,24 +4,28 @@ import android.app.Activity;
 import android.content.Intent;
 import android.media.projection.MediaProjectionManager;
 import android.os.Bundle;
+import android.widget.Spinner;
 import android.widget.Switch;
 import butterknife.ButterKnife;
 import butterknife.InjectView;
 import butterknife.OnCheckedChanged;
 import butterknife.OnClick;
-import com.jakewharton.telecine.R;
-import com.jakewharton.telecine.TelecineService;
+import butterknife.OnItemSelected;
 import javax.inject.Inject;
 import timber.log.Timber;
 
 public final class TelecineActivity extends Activity {
   private static final int CREATE_SCREEN_CAPTURE = 4242;
 
+  @InjectView(R.id.spinner_video_size_percentage) Spinner videoSizePercentageView;
   @InjectView(R.id.switch_show_countdown) Switch showCountdownView;
   @InjectView(R.id.switch_hide_from_recents) Switch hideFromRecentsView;
 
+  @Inject @VideoSizePercentage IntPreference videoSizePreference;
   @Inject @ShowCountdown BooleanPreference showCountdownPreference;
   @Inject @HideFromRecents BooleanPreference hideFromRecentsPreference;
+
+  private VideoSizePercentageAdapter videoSizePercentageAdapter;
 
   @Override protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
@@ -30,6 +34,12 @@ public final class TelecineActivity extends Activity {
 
     setContentView(R.layout.activity_main);
     ButterKnife.inject(this);
+
+    videoSizePercentageAdapter = new VideoSizePercentageAdapter(this);
+
+    videoSizePercentageView.setAdapter(videoSizePercentageAdapter);
+    videoSizePercentageView.setSelection(
+        VideoSizePercentageAdapter.getSelectedPosition(videoSizePreference.get()));
 
     showCountdownView.setChecked(showCountdownPreference.get());
     hideFromRecentsView.setChecked(hideFromRecentsPreference.get());
@@ -42,6 +52,13 @@ public final class TelecineActivity extends Activity {
         (MediaProjectionManager) getSystemService(MEDIA_PROJECTION_SERVICE);
     Intent intent = manager.createScreenCaptureIntent();
     startActivityForResult(intent, CREATE_SCREEN_CAPTURE);
+  }
+
+  @OnItemSelected(R.id.spinner_video_size_percentage) void onVideoSizePercentageSelected(
+      int position) {
+    int newValue = videoSizePercentageAdapter.getItem(position);
+    Timber.d("Video size percentage changing to %s%%", newValue);
+    videoSizePreference.set(newValue);
   }
 
   @OnCheckedChanged(R.id.switch_show_countdown) void onShowCountdownChanged() {